# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ControlHub SaaS Empresarial API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      responses:
        "200":
          description: Companies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # USERS
...
  /companies/{companyId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create document record
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /companies/{companyId}/documents/{documentId}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ANNOUNCEMENTS
...
  /companies/{companyId}/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get executive dashboard summary
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /companies/{companyId}/dashboard/alerts:
    get:
      operationId: getDashboardAlerts
      tags: [dashboard]
      summary: Get active alerts for dashboard
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /companies/{companyId}/dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # REPORTS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        ruc:
          type: string
          nullable: true
        logo:
          type: string
          nullable: true
        currency:
          type: string
        industry:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        activeModules:
          type: array
          items:
            type: string
        employeeCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, currency, activeModules, employeeCount, createdAt]
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        ruc:
          type: string
        currency:
          type: string
        industry:
          type: string
        address:
          type: string
        email:
          type: string
        phone:
          type: string
        activeModules:
          type: array
          items:
            type: string
      required: [name, currency]
    UpdateCompanyBody:
      type: object
      properties:
        name:
          type: string
        ruc:
          type: string
        currency:
          type: string
        industry:
          type: string
        address:
          type: string
        email:
          type: string
        phone:
          type: string
        activeModules:
          type: array
          items:
            type: string
...
    Document:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        employeeId:
          type: integer
          nullable: true
        employeeName:
          type: string
          nullable: true
        title:
          type: string
        category:
          type: string
          enum: [hr, finance, legal, operations, security, contracts, payslips, invoices, manuals, other]
        fileUrl:
          type: string
          nullable: true
        fileType:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        uploadedBy:
          type: string
        createdAt:
          type: string
      required: [id, companyId, title, category, uploadedBy, createdAt]
    CreateDocumentBody:
      type: object
      properties:
        employeeId:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [hr, finance, legal, operations, security, contracts, payslips, invoices, manuals, other]
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        notes:
          type: string
      required: [title, category]
...
    DashboardSummary:
      type: object
      properties:
        finance:
          type: object
          properties:
            totalPending:
              type: number
            totalOverdue:
              type: number
            totalPaid:
              type: number
            invoiceCount:
              type: integer
            overdueCount:
              type: integer
            upcomingCount:
              type: integer
          required: [totalPending, totalOverdue, totalPaid, invoiceCount, overdueCount, upcomingCount]
        hr:
          type: object
          properties:
            totalEmployees:
              type: integer
            activeEmployees:
              type: integer
            onLeaveEmployees:
              type: integer
            newThisMonth:
              type: integer
          required: [totalEmployees, activeEmployees, onLeaveEmployees, newThisMonth]
        attendance:
          type: object
          properties:
            presentToday:
              type: integer
            absentToday:
              type: integer
            lateToday:
              type: integer
            attendanceRate:
              type: number
          required: [presentToday, absentToday, lateToday, attendanceRate]
        documents:
          type: object
          properties:
            totalDocuments:
              type: integer
            recentUploads:
              type: integer
          required: [totalDocuments, recentUploads]
        announcements:
          type: object
          properties:
            totalActive:
              type: integer
            unreadCount:
              type: integer
            urgentCount:
              type: integer
          required: [totalActive, unreadCount, urgentCount]
        invoiceTrend:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              paid:
                type: number
              pending:
                type: number
              overdue:
                type: number
            required: [month, paid, pending, overdue]
      required: [finance, hr, attendance, documents, announcements, invoiceTrend]
    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [invoice_overdue, invoice_due_soon, attendance_alert, document_expiring, general]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, error]
        entityId:
          type: integer
          nullable: true
        entityType:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, type, title, message, severity, createdAt]
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        action:
          type: string
        entityName:
          type: string
        performedBy:
          type: string
        module:
          type: string
          enum: [finance, hr, attendance, documents, announcements, settings]
        createdAt:
          type: string
      required: [id, type, action, entityName, performedBy, module, createdAt]
...