# lib/api-zod/src/generated/types/createCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ControlHub SaaS Empresarial API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCompanyBody {
  name: string;
  ruc?: string;
  currency: string;
  industry?: string;
  address?: string;
  email?: string;
  phone?: string;
  activeModules?: string[];
}